--- participantes.txt
@@ -1,2 +1,0 @@
-Héctor
-Leandro
