+++ src/main/java/eggporIzquierda/solucionesactivas/controladores/ControladorServicioOfrecido.java
@@ -32,12 +32,13 @@ public class ControladorServicioOfrecido {
 
         try {
             servOfrecidoServicio.registrarServicio(serv_descripcion);
-            return "redirect:/admin";
+            return "redirect:/admin/dashboard";
         } catch (MiException ex) {
             modelo.put("error", ex.getMessage());
+            return "servicio_ofrecido_alta.html";
         }
 
-        return "servicio_ofrecido_alta.html";
+        
     }
 
     //-----------listar
