+++ src/main/java/eggporIzquierda/solucionesactivas/entity/ContratoProveedor.java
@@ -2,10 +2,12 @@ package eggporIzquierda.solucionesactivas.entity;
 
 import eggporIzquierda.solucionesactivas.enumation.EnumCalificacion;
 import eggporIzquierda.solucionesactivas.enumation.EnumEstadoContrato;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Temporal;
@@ -46,6 +48,15 @@ public class ContratoProveedor {
 
     private BigDecimal precio;
 
+
+   /*  @GeneratedValue(strategy = GenerationType.IDENTITY) 
+    @Column(nullable = false, unique = true)  
+    private Integer numeroContrato; */
+
+    /*@GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(nullable = false, unique = true)
+    private int atributoAutoincremental; */
+
 // CONSTRUCTORES:
     public ContratoProveedor() {
     }
@@ -61,6 +72,8 @@ public class ContratoProveedor {
         this.comentarioFinal = comentarioFinal;
         this.comentarioInicial = comentarioInicial;
         this.precio = precio;
+        /* this.numeroContrato = numeroContrato; */
+    
     }
 
     //GETTERS AND SETTERS
@@ -143,5 +156,11 @@ public class ContratoProveedor {
     public void setComentarioInicial(String comentarioInicial) {
         this.comentarioInicial = comentarioInicial;
     }
+
+    /*  public Integer getNumeroContrato() {
+        return numeroContrato;
+    } */
+
+   
 
 }
