+++ src/main/resources/templates/contratos_solicitados.html
@@ -15,7 +15,7 @@ Click nbfs://nbhost/SystemFileSystem/Templates/Other/html.html to edit this temp
             <div class="card" style="width: 18rem;" th:each="contrato : ${contratos}">
                 <div class="card-body">
                     <h6>N° de Contrato:</h6>
-                    <h5 class="card-title" th:text="${contrato.id}"></h5>
+                    <h5 class="card-title" th:text="${contrato.numeroContrato}"></h5>
                     <!--<img id="fotoPerfil" th:src="@{/imagen/perfil/__${usuario.id}__}" alt="imegen perfil" class="rounded-circle" width="100" height="100">-->
                     <h6>ESTADO:</h6>
                     <h6 class="card-subtitle mb-2 text-muted" th:text="${contrato.estado}"></h6>
