--- src/main/resources/templates/dashboard_admin.html
@@ -1,174 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
-
-    <head th:replace="~{/fragments/head :: head}"></head>
-
-
-    <body class="footer-alpiso d-flex flex-column" >
-
-        <header th:replace="~{/fragments/header :: header}"></header>
-
-        <main class="flex-grow-1 container-fluid">
-            <div class="container container-fluid d-flex">  
-                <div class="contenedor">
-                    <div class="rigth-panel">
-                        <img src="https://img.freepik.com/free-vector/site-stats-concept-illustration_114360-1434.jpg"  style="border-radius: 5px" width="100%" alt="Background Image">
-                    </div>
-
-                    <div class="left-panel">
-                        
-                        
-                           
-                      
-                        <nav class="sidebar sidebar-menu">
-                            <ul>
-                                <li> <h2>Panel de Administrador</h2></li>
-                                <li><a href="#">Tablero</a></li>
-                                <li><a href="#">Usuarios</a></li>
-                                <li><a href="#">Proveedores</a></li>
-                                <li><a href="#">Solicitudes de Servicios</a></li>
-                                <li><a href="#">Estadísticas</a></li>
-                                <li><a href="#">Configuraciones</a></li>
-                            </ul>
-                        </nav>
-
-                             
-
-
-
-
-                    </div>
-                </div>
-            </div>
-        </main>
-
-        <footer th:replace="~{/fragments/footer :: footer}"></footer>
-
-        <script src="https://kit.fontawesome.com/183f2a4aef.js" crossorigin="anonymous"></script>
-
-    </body>
-
-</html>
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-<!--
-
-<div class="container-fluid m-5 contenedor">
-
-    <form class="formulario" id="inicio" th:action="@{/registro}" method="POST" enctype="multipart/form-data">
-        <div class="form-group my-3">
-
-                            <h4 class="card-title">Informacion Personal</h4>
-                            <label class="mb-2"> ¿AUN NO TENES CUENTA? REGISTRATE !!!</label>
-                            <label class="mb-2"> INFORMACION PERSONAL</label>
-            <input type="text" class="form-control"  name="nombre" placeholder="Nombre" th:value="${nombre}">
-            <input type="text" class="form-control"  name="apellido" placeholder="Apellido" th:value="${apellido}">
-            <input type="text" class="form-control"  name="dni" placeholder="DNI" th:value="${dni}">
-
-            <input type="text" class="form-control"  name="telefono" placeholder="Telefono" th:value="${telefono}">
-
-                            <input type="Date" class="form-control"  name="fechaNacimiento" placeholder="Fecha de Nacimiento" th:value="${fechaNacimiento}">
-            <br>
-
-                           <h4 class="card-title">Domicilio</h4>
-                            <div class="direccion">
-                                <input type="text" class="form-control" name="calle" placeholder="Calle">
-                                <input type="text" class="form-control" name="numero" placeholder="N&uacute;mero">
-                            </div>
-                            <div class="direccion">
-                                <input type="text" class="form-control" name="barrio" placeholder="Barrio">
-                                <input type="text" class="form-control" name="lote" placeholder="Lote">
-                                <input type="text" class="form-control" name="cp" placeholder="CP">
-                            </div>
-                            <input type="text" class="form-control" name="localidad" placeholder="Localidad">
-
-        </div>
-
-        <div class="form-group my-3">
-            <input type="file" class="form-control" name="archivo">
-        </div>
-
-
-        <div class="form-group my-3">
-            <h4 class="card-title">Informacion de Inicio de Sesion!</h4>
-        </div>
-
-        <div class="form-group my-3">
-            <input type="email" class="form-control" name="email" placeholder="EMAIL" th:value="${email}">
-        </div>
-
-
-        <div class="form-group my-3">
-            <input type="password" class="form-control"  name="password" placeholder="CONTRASE&Ntilde;A"
-                   th:value="${password}">
-        </div>
-        <div class="form-group my-3">
-            <input type="password" class="form-control"  name="password2" placeholder="REPETIR CONTRASE&Ntilde;A"
-                   th:value="${password2}">
-        </div>
-
-
-        <button type="submit" class="btn btn-primary mt-3 botones">CREAR CUENTA</button>
-    </form>
-
-    <div th:if="${exito}!=null" class="card text-white bg-success mb-3 mt-3 mensajeExito">
-        <div class="card-body">
-            <h5 class="card-title">Exito!</h5>
-            <p class="card-text" th:text="${exito}"></p>
-        </div>
-    </div>
-    <div th:if="${error}!=null" class="card text-white bg-danger mb-3 mensajeError">
-        <div class="card-body">
-            <h5 class="card-title">Error 👎</h5>
-            <p class="card-text" th:text="${error}"></p>
-        </div>
-    </div>
-    <section class=" text-center container">
-        <div class="row mt-3">
-            <div class="divVolver">
-                <a th:href="@{/}" class="btn btn-secondary my-2 botonVolver botones">Volver</a>
-            </div>
-        </div>
-    </section>
-</div>
-
-<footer class="text-muted py-5">
-    <div class="containerFooter">
-        <p class="mb-1">Album example is &copy; Bootstrap,<br>but please download and customize it for yourself!</p>
-        <p class="mb-0">New to Bootstrap? <a href="/"><br>Visit the homepage</a> or read our <a
-                href="../getting-started/introduction/">getting started guide</a>.</p>
-        <p class="float-end mb-1">
-            <a href="#" style="text-decoration:none"><span>⬆️</span> Back to top</a>
-        </p>
-    </div>
-</footer>
-
-
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-pprn3073KE6tl6bjs2QrFaJGz5/SUsLqktiwsUTF55Jfv3qYSDhgCecCxMW52nD2"
-crossorigin="anonymous"></script>
-</body>
-
-</html>-->
