+++ src/main/resources/templates/fragments/header.html
@@ -26,7 +26,7 @@
             <div class="dropdown text-end" th:if="${session.usuariosession != null}">
                 <a class="d-block link-body-emphasis text-decoration-none dropdown-toggle" data-bs-toggle="dropdown"
                     aria-expanded="false">
-                    <img id="fotoPerfil" th:src="@{/imagen/perfil/__${session.usuariosession.id}__}" alt="imegen perfil"
+                    <img id="fotoPerfil" th:src="@{/imagen/perfil/__${session.usuariosession.id}__}" alt="Foto de perfil" 
                         class="rounded-circle" width="50" height="50">
                 </a>
                 <ul class="dropdown-menu text-small">
