+++ src/main/resources/templates/inicio.html
@@ -11,6 +11,17 @@
         <main class="flex-grow-1 container-fluid d-flex flex-column align-items-center">
             <h1 style="color: azure; text-align: center; margin-top: 50px;">SELECCIONA EL SERVICIO BUSCADO</h1>
 
+            <div style="text-align: center; margin-top: 20px">
+                <a th:href="@{/usuario/usuarios}">
+                    <button type="button" class="btn btn-info">Lista de Usuarios</button>
+                </a>
+                <a th:href="@{/proveedor/proveedores}">
+                    <button type="button" class="btn btn-info">
+                        Lista de Proveedores
+                    </button>
+                </a>
+            </div>
+
             <div class="d-flex flex-wrap justify-content-center" style="margin-top: 20px; text-align: center;">
                 <div class="black bg-white form-border" style="border: solid 1px; border-radius: 5px; width: 200px; height: 200px; padding: 5px; margin: 5px;">
                     <img src="https://s3.us-east-2.amazonaws.com/static.homesolution.net/res/img/cats/c-38.svg" width="140" alt="Construcción y Arquitectura" style="display: block; margin: 0 auto; align-items: center;">
