--- src/main/resources/templates/usuario_modificar.html
@@ -1,58 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
-    <head th:replace="~{/fragments/head :: head}">    </head>
-
-
-    <body>
-        <header th:replace= "~{/fragments/header :: header}"> </header>
-
-
-
-        <div class="container container-fluid d-flex">
-            <main class="flex-grow-1 container-fluid" id="mainindex">
-
-                <form class="form-control formulario" th:action="@{/usuario/usuario_modificar/__${usuario.id}__}" method="POST" enctype="multipart/form-data" >
-
-                    <input hidden th:value="${usuario.id}" name ="id">
-
-                    <h3 class="card-title" style="color: #2a363b"> Modificar datos perfil </h3>
-                    <img class="card-subtitle mb-2 text-muted rounded-circle" width="50" height="50" th:src="@{/imagen/perfil/__${usuario.id}__}"></img>
-
-
-                    <input type="text" class="form-control" name="nombre"  th:value="${usuario.nombre}">
-                    <input type="text" class="form-control" name="apellido"  th:value="${usuario.apellido}">        
-                    <input type="email" class="form-control" name="email"  th:value="${usuario.email}">
-                    <input type="text" class="form-control" name="dni" placeholder="dni" th:value="${usuario.dni}">
-                    <input type="text" class="form-control" name="telefono" placeholder="telefono" th:value="${usuario.telefono}">
-                    <h4 class="card-title" style="color: #2a363b"> Valide sus claves para guardar los cambios </h4>
-                    <input type="password" class="form-control" name="password" placeholder="CONTRASEÑA" th:value="${usuario.password}">
-                    <input type="password" class="form-control" name="password2" placeholder="REPETIR CONTRASEÑA">
-                     <input type="file" class="form-control mb-3" name="serv_imagen">
-                    <div>
-                        
-                        <h4 class="card-title" style="color: #2a363b"> Recuerde Cerrar Sesion y volver a entrar para visualizar los cambios </h4>
-                        
-                        <button type="submit" class="btn btn-primary p-2">Guardar cambios</button>
-                        <a th:href="@{/}"><button type="button" class="btn btn-primary"> Volver </button> </a>
-                    </div>
-
-                </form>
-
-            </main>
-        </div>
-
-        <footer th:replace="~{/fragments/footer :: footer}">    </footer>
-
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/js/bootstrap.bundle.min.js" integrity="sha384-pprn3073KE6tl6bjs2QrFaJGz5/SUsLqktiwsUTF55Jfv3qYSDhgCecCxMW52nD2" crossorigin="anonymous"></script>
-    </body>
-</html>
-
-
-
-
-
-
-
-<!--                        <h6>Imagen</h6>
-                        <img class="card-subtitle mb-2 text-muted rounded-circle" width="50" height="50" th:src="@{/imagen/perfil/__${usuario.id}__}"></img>-->
-
